+++ CHANGELOG.rst
@@ -14,6 +14,9 @@ Breaking changes:
 
 Improvements and new features:
 
+- Add support for creating podman pods for testing via the
+  :py:class:`~pytest_container.pod.Pod` class.
+
 - Add support for automatically exposing ports in containers via the
   :py:attr:`~pytest_container.container.ContainerBase.forwarded_ports`
   attribute: Container Images can now define which ports they want to publish